chore: Update Agencia::GraficarAvionesDisponibles method to generate graph of available airplanes
This commit updates the Agencia::GraficarAvionesDisponibles method in the Agencia class to generate a graph of all the available airplanes. The graph includes information such as flight number, registration number, and current status. This feature enhances the functionality of the project by providing a visual representation of the available airplanes. The generated graph is saved as a DOT file and converted to a PNG image for easy visualization.